--- compose-multiplatform/lib/src/commonMain/kotlin/Platform.kt
@@ -1,5 +1,0 @@
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
