--- compose-multiplatform/lib/src/commonMain/kotlin/calendar/core/CalendarDay.kt
@@ -1,13 +1,0 @@
-package calendar.core
-
-import androidx.compose.runtime.Immutable
-import kotlinx.datetime.LocalDate
-
-/**
- * Represents a day on the calendar.
- *
- * @param date the date for this day.
- * @param position the [DayPosition] for this day.
- */
-@Immutable
-data class CalendarDay(val date: LocalDate, val position: DayPosition)
