--- compose-multiplatform/lib/src/commonMain/kotlin/calendar/core/WeekDay.kt
@@ -1,13 +1,0 @@
-package calendar.core
-
-import androidx.compose.runtime.Immutable
-import kotlinx.datetime.LocalDate
-
-/**
- * Represents a day on the week calendar.
- *
- * @param date the date for this day.
- * @param position the [WeekDayPosition] for this day.
- */
-@Immutable
-data class WeekDay(val date: LocalDate, val position: WeekDayPosition)
