--- compose-multiplatform/lib/src/commonMain/kotlin/calendar/data/Utils.kt
@@ -1,16 +1,0 @@
-package calendar.data
-
-import calendar.core.YearMonth
-import kotlinx.datetime.LocalDate
-
-fun checkDateRange(startMonth: YearMonth, endMonth: YearMonth) {
-    check(endMonth >= startMonth) {
-        "startMonth: $startMonth is greater than endMonth: $endMonth"
-    }
-}
-
-fun checkDateRange(startDate: LocalDate, endDate: LocalDate) {
-    check(endDate >= startDate) {
-        "startDate: $startDate is greater than endDate: $endDate"
-    }
-}
