--- compose-multiplatform/lib/src/commonMain/kotlin/calendar/data/VisibleItemState.kt
@@ -1,9 +1,0 @@
-package calendar.data
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-internal class VisibleItemState(
-    val firstVisibleItemIndex: Int = 0,
-    val firstVisibleItemScrollOffset: Int = 0,
-)
