--- compose-multiplatform/lib/src/desktopMain/kotlin/Platform.jvm.kt
@@ -1,5 +1,0 @@
-class JVMPlatform : Platform {
-    override val name: String = "Java ${System.getProperty("java.version")}"
-}
-
-actual fun getPlatform(): Platform = JVMPlatform()
