--- compose-multiplatform/lib/src/iosMain/kotlin/Platform.ios.kt
@@ -1,7 +1,0 @@
-import platform.UIKit.UIDevice
-
-class IOSPlatform : Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
