--- compose-multiplatform/lib/src/wasmJsMain/kotlin/Platform.wasmJs.kt
@@ -1,5 +1,0 @@
-class WasmPlatform : Platform {
-    override val name: String = "Web with Kotlin/Wasm"
-}
-
-actual fun getPlatform(): Platform = WasmPlatform()
