+++ compose-multiplatform/library/build.gradle.kts
@@ -1,38 +1,28 @@
 
 import com.kizitonwose.calendar.buildsrc.Android
 import com.kizitonwose.calendar.buildsrc.Config
-import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
 import org.jetbrains.kotlin.gradle.targets.js.dsl.ExperimentalWasmDsl
-import org.jetbrains.kotlin.gradle.targets.js.webpack.KotlinWebpackConfig
 
 plugins {
     alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.androidApplication)
+    alias(libs.plugins.androidLibrary)
     alias(libs.plugins.jetbrainsCompose)
     alias(libs.plugins.composeCompiler)
+    alias(libs.plugins.mavenPublish)
 }
 
 kotlin {
     @OptIn(ExperimentalWasmDsl::class)
     wasmJs {
-        moduleName = "composeApp"
-        browser {
-            commonWebpackConfig {
-                outputFileName = "composeApp.js"
-                devServer = (devServer ?: KotlinWebpackConfig.DevServer()).apply {
-                    static = (static ?: mutableListOf()).apply {
-                        // Serve sources to debug inside browser
-                        add(project.projectDir.path)
-                    }
-                }
-            }
-        }
-        binaries.executable()
+        moduleName = "calendar"
+        browser {}
+        binaries.library()
     }
 
     androidTarget {}
 
-    jvm("desktop")
+    jvm("jvm")  // jvm("desktop")
 
     listOf(
         iosX64(),
@@ -45,25 +35,39 @@ kotlin {
         }
     }
 
-    sourceSets {
-        val desktopMain by getting
+    applyDefaultHierarchyTemplate()
 
+    sourceSets {
+        val jvmMain by getting
+        val commonMain by getting
+        val wasmJsMain by getting
+        val nativeMain by getting
+
+        androidMain.get().dependsOn(jvmMain)
         androidMain.dependencies {
             implementation(compose.preview)
-            implementation(libs.compose.activity)
         }
         commonMain.dependencies {
             implementation(compose.runtime)
             implementation(compose.foundation)
-            implementation(compose.material3)
             implementation(compose.ui)
             implementation(compose.components.resources)
             implementation(compose.components.uiToolingPreview)
-            implementation(project(":compose-multiplatform:library"))
+            api(libs.kotlinx.datetime)
         }
-        desktopMain.dependencies {
-            implementation(compose.desktop.currentOs)
+        val nonJvmMain by creating {
+            dependsOn(commonMain)
+            nativeMain.dependsOn(this)
+            wasmJsMain.dependsOn(this)
+            dependencies {}
         }
+        jvmMain.dependencies {
+//            implementation(compose.desktop.currentOs)
+        }
+    }
+    @OptIn(ExperimentalKotlinGradlePluginApi::class)
+    compilerOptions {
+        freeCompilerArgs.add("-Xexpect-actual-classes")
     }
 }
 
@@ -76,20 +80,11 @@ android {
     sourceSets["main"].resources.srcDirs("src/commonMain/resources")
 
     defaultConfig {
-        applicationId = "com.kizitonwose.calendarx"
-        minSdk = Android.minSdkSampleApp
-        targetSdk = Android.targetSdk
-        versionCode = 1
-        versionName = "1.0"
+        minSdk = Android.minSdkComposeLibrary
     }
     packaging {
         resources {
             excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-    buildTypes {
-        getByName("release") {
-            isMinifyEnabled = false
         }
     }
     java {
@@ -107,17 +102,5 @@ android {
     }
     dependencies {
         debugImplementation(compose.uiTooling)
-    }
-}
-
-compose.desktop {
-    application {
-        mainClass = "Calendar Sample"
-
-        nativeDistributions {
-            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
-            packageName = "com.kizitonwose.calendar"
-            packageVersion = "1.0.0"
-        }
     }
 }
