+++ compose-multiplatform/library/src/androidMain/AndroidManifest.xml
