+++ compose-multiplatform/library/src/commonMain/kotlin/com/kizitonwose/calendar/compose/CalendarInfo.kt
@@ -1,0 +1,12 @@
+package com.kizitonwose.calendar.compose
+
+import androidx.compose.runtime.Immutable
+import com.kizitonwose.calendar.core.OutDateStyle
+import kotlinx.datetime.DayOfWeek
+
+@Immutable
+internal data class CalendarInfo(
+    val indexCount: Int,
+    private val firstDayOfWeek: DayOfWeek? = null,
+    private val outDateStyle: OutDateStyle? = null,
+)
