+++ compose-multiplatform/library/src/commonMain/kotlin/com/kizitonwose/calendar/core/JvmSerializable.kt
@@ -1,0 +1,3 @@
+package com.kizitonwose.calendar.core
+
+expect interface JvmSerializable
