+++ compose-multiplatform/library/src/commonMain/kotlin/com/kizitonwose/calendar/core/WeekDay.kt
@@ -1,0 +1,13 @@
+package com.kizitonwose.calendar.core
+
+import androidx.compose.runtime.Immutable
+import kotlinx.datetime.LocalDate
+
+/**
+ * Represents a day on the week calendar.
+ *
+ * @param date the date for this day.
+ * @param position the [WeekDayPosition] for this day.
+ */
+@Immutable
+data class WeekDay(val date: LocalDate, val position: WeekDayPosition) : JvmSerializable
