+++ compose-multiplatform/library/src/commonMain/kotlin/com/kizitonwose/calendar/data/Utils.kt
@@ -1,0 +1,16 @@
+package com.kizitonwose.calendar.data
+
+import com.kizitonwose.calendar.core.YearMonth
+import kotlinx.datetime.LocalDate
+
+fun checkDateRange(startMonth: YearMonth, endMonth: YearMonth) {
+    check(endMonth >= startMonth) {
+        "startMonth: $startMonth is greater than endMonth: $endMonth"
+    }
+}
+
+fun checkDateRange(startDate: LocalDate, endDate: LocalDate) {
+    check(endDate >= startDate) {
+        "startDate: $startDate is greater than endDate: $endDate"
+    }
+}
