+++ compose-multiplatform/library/src/commonMain/kotlin/com/kizitonwose/calendar/data/VisibleItemState.kt
@@ -1,0 +1,9 @@
+package com.kizitonwose.calendar.data
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+internal class VisibleItemState(
+    val firstVisibleItemIndex: Int = 0,
+    val firstVisibleItemScrollOffset: Int = 0,
+)
