+++ compose-multiplatform/library/src/jvmMain/kotlin/com/kizitonwose/calendar/core/Converters.kt
@@ -1,0 +1,5 @@
+package com.kizitonwose.calendar.core
+
+import java.time.YearMonth as jtYearMonth
+
+fun YearMonth.toJavaYearMonth(): jtYearMonth = jtYearMonth.of(year, month)
