+++ compose-multiplatform/library/src/jvmMain/kotlin/com/kizitonwose/calendar/core/JvmSerializable.jvm.kt
@@ -1,0 +1,3 @@
+package com.kizitonwose.calendar.core
+
+actual typealias JvmSerializable = java.io.Serializable
