+++ compose-multiplatform/library/src/nonJvmMain/kotlin/com/kizitonwose/calendar/core/JvmSerializable.nonJvm.kt
@@ -1,0 +1,3 @@
+package com.kizitonwose.calendar.core
+
+actual interface JvmSerializable
