+++ compose-multiplatform/sample/src/commonMain/kotlin/SimpleCalendarTitle.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.text.intl.Locale
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import calendar.core.YearMonth
+import com.kizitonwose.calendar.core.YearMonth
 
 @Composable
 fun SimpleCalendarTitle(
