+++ settings.gradle.kts
@@ -18,5 +18,5 @@ include(":data")
 include(":view")
 include(":compose")
 include(":sample")
-include(":compose-multiplatform:lib")
+include(":compose-multiplatform:library")
 include(":compose-multiplatform:sample")
